pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "colonytracking"
version = "1.0.0"
description = "Image analysis tool for bacterial colony detection and tracking from time-lapse microscopy"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Biodynamics Team", email = "info@biodynamics.com"}
]
keywords = ["image-processing", "colony-tracking", "microscopy", "bacteria", "growth-analysis"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Image Recognition",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "numpy>=1.21.0",
    "scipy>=1.7.0",
    "scikit-image>=0.19.0",
    "opencv-python>=4.5.0",
    "matplotlib>=3.5.0",
    "pandas>=1.3.0",
    "Pillow>=9.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=22.0",
    "flake8>=4.0",
    "mypy>=0.950",
    "sphinx>=4.5",
]
gui = [
    "PyQt6>=6.3.0",
]

[project.urls]
Homepage = "https://github.com/biodynamics/colonytracking"
Documentation = "https://github.com/biodynamics/colonytracking/docs"
Repository = "https://github.com/biodynamics/colonytracking"
Issues = "https://github.com/biodynamics/colonytracking/issues"

[project.scripts]
colonytracking = "colonytracking.gui.main:main"

[tool.setuptools]
packages = ["colonytracking"]
package-dir = {"" = "src"}

[tool.setuptools.package-data]
colonytracking = ["assets/*", "config/*"]

[tool.black]
line-length = 100
target-version = ['py39']

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
